Add support to to(device) op. (#56857)
Currently, if we call tensor.to() method and pass a device as the parameter. It will fail, because in symbolic function of to() we didn't handle such case.

So add a check in the beginning of this symbolic function, if this is a device cast, we return self directly. A test has also been added.

Co-authored-by: Jay Zhang <jiz@microsoft.com>

ghstack-source-id: e5a436f017d752eb96c41ce4da700682992923d3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57599