[DDP] Support not all outputs used in loss calculation
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57081

Changes in this diff:
1. Enable passthrough autograd function for all cases, not just static.
2. With above, move `prepare_for_backward` which does unused parameter checking logic to beginning of backwards pass
3. Enhance process of unused parameter checking to account for outputs not being used in loss.

The way (3) is implemented is by looking at the gradient outputs at the beginning of backwards. If the gradient output corresponding to a model output is zero it means that it was not involved in loss computation (it could also mean that the gradient was indeed all zeros, but this is practically equivalent). In this case, we get the corresponding model output and mark that parameter as unused (which is equivalent to a zero gradient being accumulated).

Tested by ensuring that when a model output is not used in loss, the corresponding grad is not modified. Also verified that the grads are the same in local vs DDP training case.
ghstack-source-id: 128741360

Differential Revision: [D28048628](https://our.internmc.facebook.com/intern/diff/D28048628/)